+++ src/Events/AnalyticsOptions.ts
@@ -1,15 +1,16 @@
+import { Video } from '../types/Video';
 import { noop } from '../utils';
-import { PlaybackEvent } from './AnalyticsEvents';
 
 export interface AnalyticsOptions {
   metadata: { [key: string]: any };
-  /**
-   * @deprecated Callback signature will be replaced in a future version
-   */
-  handleOnPlayback: (event: PlaybackEvent) => void;
+  handleOnSegmentPlayback: (
+    video: Video,
+    startSeconds: number,
+    endSeconds: number,
+  ) => void;
 }
 
 export const defaultOptions: AnalyticsOptions = Object.freeze({
   metadata: {},
-  handleOnPlayback: noop,
+  handleOnSegmentPlayback: noop,
 });