+++ src/utils/convertPlaybackResource.test.ts
@@ -15,6 +15,7 @@ it('Can map the response to a StreamPlayback object', () => {
   expect(playback.id).toEqual('stream-playback-id');
   expect(playback.type).toEqual('STREAM');
   expect(playback.streamUrl).toEqual('kaltura/stream.mp4');
+  expect(playback.duration).toEqual(62);
   expect(playback.thumbnailUrl).toEqual('kaltura/poster.jpg');
   expect(playback.links.createPlaybackEvent.getOriginalLink()).toEqual(
     'create/playback/event',
@@ -29,6 +30,7 @@ it('Can map the response to a YoutubePlayback object', () => {
   expect(playback.id).toEqual('youtube-playback-id');
   expect(playback.type).toEqual('YOUTUBE');
   expect(playback.thumbnailUrl).toEqual('youtube/poster.jpg');
+  expect(playback.duration).toEqual(63);
   expect(playback.links.createPlaybackEvent.getOriginalLink()).toEqual(
     'create/playback/event',
   );